requirements.txt
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==22.2.0
blinker==1.5
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
entrypoints==0.4
fastai==2.7.11
fastcore==1.5.28
fastdownload==0.0.7
fastprogress==1.0.3
flatbuffers==23.3.3
fonttools==4.39.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
murmurhash==1.0.9
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
streamlit==1.20.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thinc==8.1.9
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.1
watchdog==2.3.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
